ci: fix to make sure gallery build function is serializable
